extract vs2015backend into own file
