+++ mesonbuild/backend/vs2010backend.py
@@ -832,21 +832,3 @@ if %%errorlevel%% neq 0 goto :VCEnd'''
         # ElementTree can not do prettyprinting so do it manually
         #doc = xml.dom.minidom.parse(ofname)
         #open(ofname, 'w').write(doc.toprettyxml())
-
-
-class Vs2015Backend(Vs2010Backend):
-    def __init__(self, build):
-        super().__init__(build)
-        self.platform_toolset = 'v140'
-
-    @staticmethod
-    def has_objects(objects, additional_objects, generated_objects):
-        # VS2015 requires generated objects to be added explicitly to the project file.
-        return len(objects) + len(additional_objects) + len(generated_objects) > 0
-
-    @staticmethod
-    def add_generated_objects(node, generated_objects):
-        # VS2015 requires generated objects to be added explicitly to the project file.
-        for s in generated_objects:
-            ET.SubElement(node, 'Object', Include=s)
-        return
