+++ mesonbuild/mesonmain.py
@@ -139,8 +139,8 @@ itself as required.'''
             from .backend import vs2010backend
             g = vs2010backend.Vs2010Backend(b)
         elif self.options.backend == 'vs2015':
-            from .backend import vs2010backend
-            g = vs2010backend.Vs2015Backend(b)
+            from .backend import vs2015backend
+            g = vs2015backend.Vs2015Backend(b)
         elif self.options.backend == 'xcode':
             from .backend import xcodebackend
             g = xcodebackend.XCodeBackend(b)
